server/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2-binary>=2.9.1,<2.9.2
djangorestframework-simplejwt==5.0.0
djoser==2.0.5
django-jazzmin==2.4.8
gunicorn==20.1.0
whitenoise==5.3.0

# AWS S3 Storage
django-storages==1.12.3
boto3==1.24.0

# Payment Processing
stripe==4.0.2

# Nested Routers for API
drf-nested-routers==0.93.4

# Facebook Marketing API
facebook-business==17.0.0

# Caching and Task Queue
redis==4.3.4
celery==5.2.7
django-redis==5.2.0

# Database Connection Pooling
django-db-connection-pool==1.0.5